Installer redesigned (read below).
From now on, the files are not included in the installer itself, so it won't have to be rebuilt every time.
It's designed to unpack 7z archives from "game" subdir, which can be prepared using prepare_archives.cmd script.
7z files are passworded! If you want to change the password, you have to do it both in the archives (prepare_archives.cmd)
and the installer itself, so it can unpack the files!
3rd party prerequisities are also as external installers now. They are called by the installer from "3rparties" subdir.

Read the LostAlpha.iss script for more information about the structure.